requirements.txt
fastapi~=0.135.3
uvicorn[standard]~=0.44.0
ollama~=0.6.1
sqlglot~=30.4.3
aiosqlite~=0.22.1
pyyaml~=6.0.3
pytest~=9.0.3
pandas~=3.0.2
pytest-asyncio